U: time.h time_from_* casting

+++ xo/time/time.h
@@ -47,9 +47,9 @@ namespace xo
 		storage_t internal_;
 	};
 
-	template< typename T > time time_from_seconds( T s ) { return time( time::storage_t( s * 1e9 + 0.5 ) ); }
-	template< typename T > time time_from_milliseconds( T s ) { return time( time::storage_t( s * 1e6 + 0.5 ) ); }
-	template< typename T > time time_from_nanoseconds( T s ) { return time( time::storage_t( s + 0.5 ) ); }
+	template< typename T > time time_from_seconds( T s ) { return time( time::storage_t( double( s ) * 1e9 + 0.5 ) ); }
+	template< typename T > time time_from_milliseconds( T s ) { return time( time::storage_t( double( s ) * 1e6 + 0.5 ) ); }
+	template< typename T > time time_from_nanoseconds( T s ) { return time( time::storage_t( double( s ) + 0.5 ) ); }
 
 	inline namespace literals {
 		constexpr time operator ""_ns( unsigned long long ns ) { return time( time::storage_t( ns ) ); }
